[DebugInfo] Add symbol for debugger with VTable information.
Address comments from reviewers:
- Remove the LLVM tests, as they do not add any coverage.
- Rework the test case that use individual .cpp/h files to
  use just a single file. The checks are the same.
