--- clang/test/CodeGenCXX/Inputs/vtable-debug-info-inheritance-simple-derived.h
@@ -1,14 +1,0 @@
-#include "vtable-debug-info-inheritance-simple-base.h"
-
-#ifndef DERIVED_H
-#define DERIVED_H
-
-struct CDerived : NSP::CBase {
-  unsigned D = 2;
-  void zero() override;
-  int two() override;
-  int three() override;
-};
-
-extern void fooDerived();
-#endif
